requirements.txt
python-telegram-bot[job-queue]>=21.0
feedparser>=6.0.10
twikit>=1.4.1
groq>=0.9.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
requests>=2.31.0
httpx>=0.27.0
PyYAML>=6.0.1
